require domain names in frame src
By requiring the author to provide a more complete url, forks of frame
items will generally still be able to render the content as a remote
page (remote relative to the domain of the forked page).

We also save the found hostname in anticipation of addressing some
XSS security concerns.

Closes #1