Simplify iframe sandbox & remove apparatus example
Fixes #2

This changes our decision in #2.

The complexity in the sandbox attribute has made it hard to reason
about the security of the frame plugin. This solution sets the sandbox
to always be 'allow-scripts'.

In this change, we choose to accept the risk that scripts inside
assets in the same origin, when shared through a frame, will be able
to modify the DOM of the hosting wiki. This risk is mitigates somewhat
in that uploading assets requires an authenticated admin. As such, it
is really no more risky than admins installing wiki plugins via
plugmatic. Moreover, it empowers motivated programmers to experiment
with extending wiki before committing to plugins.

This change also removes the apparatus example from the about page.
Previous sandbox complexity was motivated in part to support
apparatus. We have learned that apparatus can be partially supported
with CORS headers on the apparatus hosting site.  However, as we write
at least some browsers forbid changes to localStorage inside the
frame. The apparatus diagram can be manipulated once before triggering
the security restrictions.