Fix reload loop introduced when switching to functions
This was a mess to track down. Fix ends up be quite simple.

A reload loop was introduced by removing `var location` at the top of
the file (along with all the other var declarations we removed). That
var declaration was protecting the code from the line at the bottom of
the file which did this:

  location = window.location

With `var location` at the top of the file, the above line just saves
a locally scoped variable. Removing that declaration changes the
meaning of that line to this:

  window.location = window.location

I believe the 'location' variable was introduced to support a specific
test case which was later removed or never completed.

Anyway, this commit removes the unused 'location' variable and also
catches one more function declaration we'd missed in the previous
commit.