Add LINEUP keyword to listen for streaming sources
The SOURCE and LINEUP keywords are written to play nicely
together.

A SOURCE emits events. And emits events narrowly.

A LINEUP listens for events. And listens widely.

To be more precise, SOURCE emits events from the
.item.frame DOM element. The events deliberately bubble.
LINEUP listens for the events from the .main div of the
lineup.

The SOURCE events also provide identifiers in the event
details. LINEUP listeners can be more selective in their
listening by filtering the event stream with those ids.