security fix: set sandbox before src
We discovered confusing behavior where same-origin HTML scripts had
extra privileges when first loaded. Reloading the frame would enforce
the expected privileges.

Thanks to paul90, we discovered that order matters when setting
attributes on a new iframe. The restricted behavior we want is
correctly enforced if we set the sandbox attribute first.