HTML scripts from foreign origins can now access cookies & storage
If the frame src is a different origin, then allow-same-origin is
permitted.

https://developer.mozilla.org/en-US/docs/Web/HTML/Element/iframe

    When the embedded document has the same origin as the embedding
    page, it is strongly discouraged to use both allow-scripts and
    allow-same-origin, as that lets the embedded document remove the
    sandbox attribute — making it no more secure than not using the
    sandbox attribute at all.

Therefore we forbid allow-same-origin when the frame src matches the
wiki origin.

Access to cookies and browser storage enables frames to be used for
deep links into youtube videos and also enables some ObservableHQ
examples to be imported, for example  voronoi.update
https://observablehq.com/embed/@d3/voronoi-update?cells=chart