Different return types. :-( Sandi Metz calls this untrustworthy code
Adding comments for now so I don't forget.

Here's an example where Sandi Metz talks about the problem with
untrustworthy code:
https://sandimetz.com/blog/2014/12/19/suspicions-of-nil

But to paraphrase, this code forces clients writing scripts to have to
check the type on the value we return to them. We'd like to make this
better but need to proceed carefully to avoid breaking other people's
code.