moved events tests into their own file
Over time, we have gradually accumulated many affordances for the
frame plugin. All of them are entangled with browser APIs. In the past
(few? many?) months we have come to depend on these affordances for
many innovations in federated wiki.

With this commit, we can add tests around the affordances we in
anticipation of promoting some of this code into the wiki client
itself.